requirements.txt
aiohttp
xmltodict
python-socketio>=4,<5
websocket-client
